refactor: move files to pipe-harness directory
