+++ packages/spectacular/src/index.ts
@@ -16,9 +16,9 @@ export * from './lib/feature-testing/navigation/spectacular-feature-location';
 export * from './lib/feature-testing/navigation/spectacular-feature-router';
 
 // Pipe testing API
-export * from './lib/pipe-testing/create-pipe-test-harness';
-export * from './lib/pipe-testing/create-pipe-test-harness-options';
-export * from './lib/pipe-testing/pipe-test-harness';
+export * from './lib/pipe-testing/pipe-harness/create-pipe-test-harness';
+export * from './lib/pipe-testing/pipe-harness/create-pipe-test-harness-options';
+export * from './lib/pipe-testing/pipe-harness/pipe-test-harness';
 
 // Shared testing API
 export * from './lib/shared/app-component/spectacular-app.component';
