+++ packages/spectacular/src/lib/pipe-testing/pipe-harness/pipe-test-harness.ts
